Transaction Tested & additional methods deleted.
