+++ src/main/java/use_case/TransactionDTO.java
@@ -3,6 +3,7 @@ package use_case;
 import model.Money;
 
 import java.math.BigDecimal;
+import java.util.Objects;
 
 public class TransactionDTO {
 
@@ -20,21 +21,6 @@ public class TransactionDTO {
         this.transactionType = transactionType;
     }
 
-    public String getAccountNumber() {
-        return accountNumber;
-    }
-
-    public Money getMoney() {
-        return money;
-    }
-
-    public BigDecimal getCustomerBalance() {
-        return customerBalance;
-    }
-
-    public String getTransactionType() {
-        return transactionType;
-    }
 
     @Override
     public String toString() {
@@ -44,5 +30,20 @@ public class TransactionDTO {
                 ", customerBalance=" + customerBalance +
                 ", transactionType='" + transactionType + '\'' +
                 '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        TransactionDTO that = (TransactionDTO) o;
+        return Objects.equals(accountNumber, that.accountNumber) && Objects.equals(money, that.money)
+                && Objects.equals(customerBalance, that.customerBalance)
+                && Objects.equals(transactionType, that.transactionType);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(accountNumber, money, customerBalance, transactionType);
     }
 }
