+++ src/test/java/FakeTransactions.java
@@ -19,10 +19,10 @@ public class FakeTransactions implements Transactions {
                                                , new BigDecimal(50_000),"DEPOSIT" ));
 
         transactionList.add(new TransactionDTO("1298574125",new Money(new BigDecimal(10_000), EURO)
-                                               , new BigDecimal(50_000),"DEPOSIT" ));
+                                               , new BigDecimal(50_000),"Withdraw" ));
 
         transactionList.add(new TransactionDTO("1298574125",new Money(new BigDecimal(10_000), EURO)
-                                               , new BigDecimal(50_000),"DEPOSIT" ));
+                                               , new BigDecimal(50_000),"Transfer" ));
     }
 
     @Override
